# backend/api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AapadBandhav Smart Emergency Response API",
    "version": "2.0.0",
    "description": "## AapadBandhav Platform API\n\n**AapadBandhav** is a Smart Emergency Response and Accident Detection Platform for India.\n\n### Features\n- 🚨 Real-time accident detection via IoT crash sensors\n- 🚑 Smart dispatch to hospitals, ambulances, police, fire departments\n- 🗺️ Live GPS tracking of responders and affected users\n- 📲 OTP-based authentication for all entity types\n- 🛡️ Full RBAC with 11 roles\n\n### Authentication\nAll protected endpoints require a **Bearer JWT token**.\n\nTo authenticate:\n1. Send OTP → `POST /api/auth/otp/send`\n2. Verify OTP → `POST /api/auth/otp/verify` → get **token**\n3. Click **Authorize** above → enter: `Bearer <your-token>`\n\nFor admin access use `POST /api/auth/admin/login`.\n\n### Roles\n`user` · `volunteer` · `fire_department` · `emergency_personnel`  \n`hospital` · `ambulance` · `police_station` · `policeman`  \n`mechanic` · `insurance` · `admin` · `superadmin`",
    "contact": {
      "name": "AapadBandhav Team",
      "email": "admin@aapadbandhav.in"
    },
    "license": {
      "name": "Private"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "https://aapadbandhav.vercel.app",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token. Get one from POST /api/auth/otp/verify or POST /api/auth/admin/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          }
        }
      },
...
      "EmergencyContact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Father"
          },
          "mobile": {
            "type": "string",
            "example": "9876543210"
          },
          "relation": {
            "type": "string",
            "example": "Father"
          }
        }
      },
...
    }
  },
  "security": [],
...
  "paths": {
...
    "/api/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get my profile",
        "description": "Returns the full profile of the authenticated entity (works for all 12 entity types).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Entity profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "entityType": {
                      "type": "string",
                      "example": "user"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Update my profile",
        "description": "Updates profile fields for the authenticated entity. Accepts any combination of profile fields.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "example": "Ramesh Kumar"
                  },
                  "email": {
                    "type": "string",
                    "example": "ramesh@gmail.com"
                  },
                  "address": {
                    "type": "string"
                  },
                  "bloodGroup": {
                    "type": "string",
                    "example": "O+"
                  },
                  "age": {
                    "type": "integer",
                    "example": 28
                  },
                  "gender": {
                    "type": "string",
                    "example": "Male"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 16.5062
                  },
                  "longitude": {
                    "type": "number",
                    "example": 80.648
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get citizen user profile (alias)",
        "description": "Returns profile for user/volunteer/fire_department roles. Alias for GET /api/profile.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Update citizen user profile (alias)",
        "description": "Updates user profile. Alias for PUT /api/profile.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "bloodGroup": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  },
                  "gender": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/emergency-contacts": {
      "get": {
        "tags": [
          "Emergency Contacts"
        ],
        "summary": "List emergency contacts",
        "description": "Returns all emergency contacts for the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of emergency contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "contacts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmergencyContact"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Emergency Contacts"
        ],
        "summary": "Add emergency contact",
        "description": "Creates a new emergency contact for the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "mobile",
                  "relation"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Father"
                  },
                  "mobile": {
                    "type": "string",
                    "example": "9876543210"
                  },
                  "relation": {
                    "type": "string",
                    "example": "Father"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emergency contact created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "contact": {
                      "$ref": "#/components/schemas/EmergencyContact"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/emergency-contacts/{id}": {
      "put": {
        "tags": [
          "Emergency Contacts"
        ],
        "summary": "Update emergency contact",
        "description": "Updates an existing emergency contact by ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Emergency contact ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mobile": {
                    "type": "string"
                  },
                  "relation": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Emergency Contacts"
        ],
        "summary": "Delete emergency contact",
        "description": "Deletes an emergency contact by ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Emergency contact ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}